+++ paq8pxd.cpp
@@ -143,6 +143,7 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "filters/mdffilter.hpp" 
 #include "filters/cdfilter.hpp" 
 #include "filters/giffilter.hpp" 
+#include "filters/szddfilter.hpp" 
 
 Img24Filter img24("image 24bit");
 Img32Filter img32("image 32bit");
@@ -153,6 +154,7 @@ EOLFilter    eolf("eol");
 MDFFilter    mdff("mdf");
 CDFilter    cdff("cd");
 gifFilter    giff("gif");
+szddFilter    szddf("szdd");
 
 Streams streams;
 /////////////////////// Global context /////////////////////////
@@ -297,253 +299,6 @@ inline bool is_base85(unsigned char c) {
     return (isalnum(c) || (c==13) || (c==10) || (c=='y') || (c=='z') || (c>='!' && c<='u'));
 }
 
-//LZSS compressor/decompressor class
-//http://my.execpc.com/~geezer/code/lzss.c
-class LZSS {
-    private:
-    const U32 N;                // size of ring buffer
-    const U32 F;                // upper limit for g_match_len.
-                                // 16 for compatibility with Microsoft COMPRESS.EXE and EXPAND.EXE
-    const U32 THRESHOLD;        // encode string into position and length if match_length is greater than this
-    U32  NIL;                   // index for root of binary search trees
-    Array<U8> LZringbuffer;     // ring buffer of size N, with extra F-1 bytes
-                                // to facilitate string comparison
-    U32 matchpos;               // position and length of longest match; set by insert_node()
-    U32 matchlen;
-    Array<U32> LZ_lchild;       // left & right children & parent -- these constitute binary search tree
-    Array<U32> LZ_rchild;
-    Array<U32> LZ_parent;
-    File*g_infile, *g_outfile; //input and output file to be compressed
-    U32 filesizez;
-
-    // Inserts string of length F, LZringbuffer[r..r+F-1], into one of the
-    // trees (LZringbuffer[r]'th tree) and returns the longest-match position
-    // and length via the global variables matchpos and matchlen.
-    // If matchlen = F, then removes the old node in favour of the new
-    // one, because the old one will be deleted sooner.
-    // Note r plays double role, as tree node and position in buffer.
-void insert_node(int r){
-    U8 *key;
-    U32 i, p;
-    int cmp;
-    cmp = 1;
-    key = &LZringbuffer[r];
-    p=N+1+key[0];
-    LZ_rchild[r]=LZ_lchild[r]=NIL;
-    matchlen = 0;
-    while(1){
-        if(cmp>= 0){
-            if(LZ_rchild[p]!=NIL) p=LZ_rchild[p];
-            else{
-                LZ_rchild[p]=r;
-                LZ_parent[r]=p;
-                return;
-            }
-        }
-        else{
-            if(LZ_lchild[p]!=NIL) p=LZ_lchild[p];
-            else{
-                LZ_lchild[p]=r;
-                LZ_parent[r]=p;
-                return;
-            }
-        }
-        for(i=1;i<F;i++){
-            cmp=key[i]-LZringbuffer[p+i];
-            if(cmp != 0) break;
-        }
-        if(i>matchlen){
-            matchpos=p;
-            matchlen=i;
-            if(matchlen>=F) break;
-        }
-    }
-    LZ_parent[r]=LZ_parent[p];
-    LZ_lchild[r]=LZ_lchild[p];
-    LZ_rchild[r]=LZ_rchild[p];
-    LZ_parent[LZ_lchild[p]]=r;
-    LZ_parent[LZ_rchild[p]]=r;
-    if(LZ_rchild[LZ_parent[p]]==p) LZ_rchild[LZ_parent[p]]=r;
-    else LZ_lchild[LZ_parent[p]]=r;
-    LZ_parent[p]=NIL;                   // remove p
-}
-
-//deletes node p from tree
-void delete_node(unsigned p){
-    U32 q;
-    if(LZ_parent[p]==NIL) return;       // not in tree
-    if(LZ_rchild[p]==NIL) q=LZ_lchild[p];
-    else if(LZ_lchild[p]==NIL) q=LZ_rchild[p];
-    else{
-        q=LZ_lchild[p];
-        if(LZ_rchild[q]!=NIL){
-            do q=LZ_rchild[q];
-            while(LZ_rchild[q]!=NIL);
-            LZ_rchild[LZ_parent[q]]=LZ_lchild[q];
-            LZ_parent[LZ_lchild[q]]=LZ_parent[q];
-            LZ_lchild[q]=LZ_lchild[p];
-            LZ_parent[LZ_lchild[p]]=q;
-        }
-        LZ_rchild[q]=LZ_rchild[p];
-        LZ_parent[LZ_rchild[p]]=q;
-    }
-    LZ_parent[q] = LZ_parent[p];
-    if(LZ_rchild[LZ_parent[p]]==p) LZ_rchild[LZ_parent[p]]=q;
-    else LZ_lchild[LZ_parent[p]] = q;
-    LZ_parent[p]=NIL;
-}
-public:
-    U32 usize;
-    LZSS(File*in, File* out,U32 fsize,U32 qn);
-
-//may fail when compressed size is larger the input (uncompressible data)
-U32 compress(){
-    U32 i, len, r, s, last_match_length, code_buf_ptr;
-    U8 code_buf[17], mask;
-    U32 ocount;
-    int c;
-    // code_buf[1..16] saves eight units of code, and code_buf[0] works as
-    // eight flags, "1" representing that the unit is an unencoded letter (1 byte),
-    // "0" a position-and-length pair (2 bytes). Thus, eight units require at most
-    // 16 bytes of code.
-    ocount=0;
-    code_buf[0]=0;
-    code_buf_ptr=mask=1;
-    s=0;
-    r=N-F;
-    // Clear the buffer with any character that will appear often.
-    memset(&LZringbuffer[0]+s,' ',r-s);
-    // Read F bytes into the last F bytes of the buffer
-    for(len=0;len<F;len++){
-        c=g_infile->getc();
-        if(c==EOF)break;
-        LZringbuffer[r+len]=c;
-    }
-    if(len==0) return 0; //text of size zero
-    // Insert the F strings, each of which begins with one or more 'space'
-    // characters. Note the order in which these strings are inserted.
-    // This way, degenerate trees will be less likely to occur.
-    for(i=1; i<=F;i++) insert_node(r-i);
-    // Finally, insert the whole string just read. The global variables
-    // matchlen and matchpos are set.
-    insert_node(r);
-    do{
-        // matchlen may be spuriously long near the end of text.
-        if(matchlen>len) matchlen=len;
-        if(matchlen<=THRESHOLD){            // Not long enough match. Send one byte.
-            matchlen=1;
-            code_buf[0]|=mask;              // 'send one byte' flag 
-            code_buf[code_buf_ptr]=LZringbuffer[r];  // Send uncoded.
-            code_buf_ptr++;
-        }
-        else{                               // Send position and length pair. Note matchlen > THRESHOLD.
-            code_buf[code_buf_ptr]=(U8)matchpos;
-            code_buf_ptr++;
-            code_buf[code_buf_ptr]=(U8)(((matchpos>>4)&0xF0)|(matchlen-(THRESHOLD+1)));
-            code_buf_ptr++;
-        }
-        mask<<=1;                           // Shift mask left one bit.
-        if(mask==0){                        // Send at most 8 units of code together
-            for(i=0;i<code_buf_ptr;i++){
-                g_outfile->putc(code_buf[i]),ocount++;
-                if(ocount>=filesizez) return ocount;
-            }
-            code_buf[0]=0;
-            code_buf_ptr=mask=1;
-        }
-        last_match_length=matchlen;
-        for(i=0;i<last_match_length;i++){
-            c=g_infile->getc();
-            if(c==EOF) break;
-            delete_node(s);                 // Delete old strings and read new bytes
-            LZringbuffer[s] = c;
-            // If the position is near the end of buffer, extend the buffer
-            // to make string comparison easier.
-            // Since this is a ring buffer, increment the position modulo N.
-            // Register the string in LZringbuffer[r..r+F-1] 
-            if(s<F-1) LZringbuffer[s+N]=c;
-            s=(s+1)&(N-1);
-            r=(r+1)&(N-1);
-            insert_node(r);
-        }
-        while(i++<last_match_length){       // After the end of text,
-            delete_node(s);                 // no need to read, but
-            s=(s+1)&(N-1);
-            r=(r+1)&(N-1);
-            len--;
-            if(len) insert_node(r);         // buffer may not be empty.
-        }
-    } while(len>0);                         //until length of string to be processed is zero
-    if(code_buf_ptr>1){                     // Send remaining code.
-        for(i=0;i<code_buf_ptr;i++){
-            g_outfile->putc(code_buf[i]),ocount++;
-            if(ocount>=filesizez) return ocount;
-        }
-    }
-    return ocount;    //return compressed size
-}
-
-U32 decompress(){
-    U32 r, flags;
-    int i,c, j, k;
-    U32 icount,incount;
-    icount=incount=0;
-    memset(&LZringbuffer[0],' ',N-F);
-    r = N - F;
-    for(flags=0;;flags>>=1){
-    // Get a byte. For each bit of this byte:
-    // 1=copy one byte literally, from input to output
-    // 0=get two more bytes describing length and position of previously-seen
-    // data, and copy that data from the ring buffer to output
-        if((flags&0x100)==0){
-            c=g_infile->getc(),incount++;
-            if(c==EOF||icount>=filesizez) break;
-            flags=c|0xFF00;
-        }
-        if(flags & 1){
-            c=g_infile->getc(),incount++;
-            if(c==EOF||icount>=filesizez) break;
-            g_outfile->putc(c),icount++;
-            LZringbuffer[r]=c;
-            r=(r+1)&(N-1);
-        }
-        // 0=get two more bytes describing length and position of previously-
-        // seen data, and copy that data from the ring buffer to output
-        else{
-            i=g_infile->getc(),incount++;
-            if(i==EOF||icount>=filesizez) break;
-            j=g_infile->getc(),incount++;
-            if(j==EOF ||icount>=filesizez) break;
-            i|=((j&0xF0)<< 4);
-            j=(j&0x0F)+THRESHOLD;
-            for(k=0;k<=j;k++){
-                c=LZringbuffer[(i+k)&(N-1)];
-                g_outfile->putc(c),icount++;
-                LZringbuffer[r]=c;
-                r=(r+1)&(N-1);
-            }
-        }
-    }
-    usize=icount;       //decompressed size
-    return incount-1;   //return compressed size
-}
-};
-LZSS::LZSS(File*in, File* out,U32 fsize,U32 qn=0): N(4096),F(16+qn),THRESHOLD(2),NIL(N),
-LZringbuffer(N+F-1), LZ_lchild(N+1), LZ_rchild(N+257), LZ_parent(N+1),filesizez(fsize),usize(0){
-    g_infile=in, g_outfile=out;
-    // initialize trees
-    // For i = 0 to N - 1, LZ_rchild[i] and LZ_lchild[i] will be the right and
-    // left children of node i. These nodes need not be initialized.
-    // Also, LZ_parent[i] is the parent of node i. These are initialized to
-    // NIL (= N), which stands for 'not used.'
-    // For i = 0 to 255, LZ_rchild[N + i + 1] is the root of the tree
-    // for strings that begin with character i. These are initialized
-    // to NIL.  Note there are 256 trees.
-    for(U32 i=N+1;i<=N+256;i++)
-        LZ_rchild[i]=NIL;
-    for(U32 i=0;i<N;i++)
-        LZ_parent[i]=NIL;
-}
 
 //read compressed word,dword
 U32 GetCDWord(File*f){
@@ -2298,7 +2053,7 @@ Filetype detect(File* in, U64 n, Filetype type, int &info, int &info2, int it=0)
     else{
       text.lastSpace++;
       text.spaceRun = 0;
-    }
+    } 
     text.wordLength = (text.isLetter)?text.wordLength+1:0;
     text.missCount-=text.misses>>31;
     text.misses<<=1;
@@ -4858,38 +4613,6 @@ void encode_ascii85(File* in, File* out, int len) {
     out->blockwrite(&ptr[0],   olen  );
 }
 
-//SZDD
-int decode_szdd(File*in, int size, int info, File*out, FMode mode, U64 &diffFound){
-    LZSS* lz77;
-    int r=0;
-    //Write out or compare
-    if (mode==FDECOMPRESS){
-            lz77=new LZSS(in,out,size,(info>>25)*2);
-             r=lz77->compress();
-            delete lz77;
-        }
-    else if (mode==FCOMPARE){
-        FileTmp out1;
-        lz77=new LZSS(in,&out1,size,(info>>25)*2);
-        r=lz77->compress();
-        delete lz77;
-        out1.setpos(0);
-        for(int i=0;i<r;i++){
-            U8 b=out1.getc();
-            if (b!=out->getc() && !diffFound) diffFound= out->curpos();
-        }
-        out1.close();
-    }
-    return r;
-}
-
-void encode_szdd(File* in, File* out, int len) {
-    LZSS* lz77;
-    lz77=new LZSS(in,out,len&0x1ffffff,(len>>25)*2);
-    lz77->decompress();
-    delete lz77;
-}
-
 //////////////////// Compress, Decompress ////////////////////////////
 
 //for block statistics, levels 0-5
@@ -4952,7 +4675,9 @@ void transform_encode_block(Filetype type, File*in, U64 len, Encoder &en, int in
         else if (type==BASE64) encode_base64(in, tmp, int(len));
         else if (type==UUENC) encode_uud(in, tmp, int(len),info);
         else if (type==BASE85) encode_ascii85(in, tmp, int(len));
-        else if (type==SZDD) encode_szdd(in, tmp, info);
+        else if (type==SZDD) {
+            szddf.encode(in, tmp,0, info);
+        }
         else if (type==ZLIB) diffFound=encode_zlib(in, tmp, int(len))?0:1;
         else if (type==BZIP2) encode_bzip2(in, tmp, len,info);
         else if (type==CD) cdff.encode(in, tmp, (len), info);
@@ -5414,7 +5139,10 @@ U64 decompressStreamRecursive(File*out, U64 size, Encoder& en, FMode mode, int i
                 if (type==BASE64) len=decode_base64(&tmp, int(len), out, mode, diffFound);
                 else if (type==UUENC)  len=decode_uud(&tmp, int(len), out, mode, diffFound);
                 else if (type==BASE85) len=decode_ascii85(&tmp, int(len), out, mode, diffFound);
-                else if (type==SZDD)   len=decode_szdd(&tmp,info,info ,out, mode,  diffFound);
+                else if (type==SZDD)  {
+                    diffFound=szddf.CompareFiles(&tmp,out,info,uint64_t(info),mode);
+                    len=szddf.fsize; // get decoded size
+                }
                 else if (type==ZLIB)   len=decode_zlib(&tmp,int(len),out, mode, diffFound);
                 else if (type==BZIP2)  len=decode_bzip2(&tmp,int(len),out, mode, diffFound,info);
                 else if (type==CD)     {
